+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log for ARCUtils
 
+## 2.10.0 - 2016-05-05
+
+- Corrected default ARC.cdn.paths.jquery-js setting.
+- Added ARC-specific Django Admin stuff (added top level `admin`
+  module).
+
+
 ## 2.9.1 - 2016-04-11
 
 ## Fixed